Rename all handler files
