+++ server/src/routes/account/handlers/base.handler.ts
