+++ server/src/routes/auth/handlers/base.handler.ts
