+++ server/src/routes/auth/handlers/register.handler.ts
